small refacto agents
