+++ packages/agents/src/index.ts
@@ -1,5 +1,6 @@
 export * from './types';
-export * from './config';
-export * from './ragAgentFactory';
+export * from './config/settings';
+export * from './config/agent';
+export * from './core/agentFactory';
 export * from './db/vectorStore';
-export * from './config';
+export * from './utils';
