--- packages/agents/src/utils/utils.ts
@@ -1,14 +1,0 @@
-import { BaseMessage } from '@langchain/core/messages';
-
-export const formatChatHistoryAsString = (history: BaseMessage[]) => {
-  return history
-    .map((message) => `${message._getType()}: ${message.content}`)
-    .join('\n');
-};
-
-export const parseXMLContent = (xml: string, tag: string): string[] => {
-  const regex = new RegExp(`<${tag}>(.*?)</${tag}>`, 'gs');
-  const matches = [...xml.matchAll(regex)];
-  return matches.map((match) => match[1].trim());
-};
-
