+++ packages/backend/src/server.ts
@@ -2,8 +2,8 @@ import express from 'express';
 import http from 'http';
 import cors from 'cors';
 import { initializeLLMConfig } from './config/llm';
-import { getPort } from '@starknet-agent/agents/config';
-import logger from '@starknet-agent/agents/utils/logger';
+import { getPort } from '@starknet-agent/agents/config/settings';
+import { logger } from '@starknet-agent/agents/utils/index';
 import { initializeHTTP } from './config/http';
 import { Container } from './config/context';
 import { validateConfig } from './config/validateConfig';
