+++ packages/backend/src/types/index.ts
@@ -1,7 +1,8 @@
 import eventEmitter from 'events';
 import { BaseMessage } from '@langchain/core/messages';
 import { Embeddings } from '@langchain/core/embeddings';
-import { LLMConfig, VectorStore } from '@starknet-agent/agents/index';
+import { LLMConfig } from '@starknet-agent/agents/types/index';
+import { VectorStore } from '@starknet-agent/agents/db/vectorStore';
 import { BaseChatModel } from '@langchain/core/language_models/chat_models';
 import { CorsOptions } from 'cors';
 import { Express } from 'express';
