+++ packages/ingester/src/generateEmbeddings.ts
@@ -1,10 +1,10 @@
 import dotenv from 'dotenv';
 import { createInterface } from 'readline';
 import logger from '@starknet-agent/agents/utils/logger';
-import { VectorStore } from '@starknet-agent/agents/index';
+import { VectorStore } from '@starknet-agent/agents/db/vectorStore';
 import { getVectorDbConfig } from '@starknet-agent/agents/config';
 import { loadOpenAIEmbeddingsModels } from '@starknet-agent/backend/lib/providers/openai';
-import { DocumentSource } from '@starknet-agent/agents/core/types';
+import { DocumentSource } from '@starknet-agent/agents/types/index';
 import { IngesterFactory } from './IngesterFactory';
 
 /**
